requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==4.0.10
djangorestframework==3.13.1
gunicorn==23.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pytz==2025.1
sqlparse==0.5.3
whitenoise==6.9.0
